+++ src/Components/3dModels/Closet/Section/index.jsx
@@ -3,10 +3,12 @@ import Structure from "../Structure"
 import Shelf from "../Shelf";
 import Leg from "../Leg";
 
-function Section({ dimensions, position, numOfShelves = 4, withLegs = true, oneDoor= false }) {
+function Section({ dimensions, position, numOfShelves = 4, withLegs = true, oneDoor = false }) {
     const { materials } = useGLTF('/assets/3dModels/Materials.glb')
-    const shelvesArr = Array.from({ length: numOfShelves }, (_, i) => i);
     const legGap = 0.07;
+
+
+
     return (
         <>
             <group position={[position.X, position.Y, position.Z]}>
@@ -25,30 +27,10 @@ function Section({ dimensions, position, numOfShelves = 4, withLegs = true, oneD
                         Z: 0
                     }}
                     material={materials.wood}
+                    numOfShelves = {numOfShelves}
+                    
                 />
 
-
-                {/* creating the shelves: */}
-
-                {shelvesArr.map((i, n) => {
-                    return (
-                        <Shelf
-                            key={n}
-                            dimensions={dimensions}
-                            material={materials.wood}
-                            position={{
-                                X: dimensions.X / 2,
-                                Y: (
-                                    -(dimensions.Y / 2) + //starting point
-                                    (dimensions.Y / (numOfShelves + 1)) + //skip the bottom
-                                    ((dimensions.Y - (dimensions.Y / (numOfShelves + 1))) //skip the top
-                                        / numOfShelves) * i //add height for each shelf in the array
-
-                                ),
-                                Z: 0
-                            }} />
-                    )
-                })}
