+++ src/Data/fakeDataTest.js
@@ -123,8 +123,8 @@ const fakeOrderData = {
         drawers: 2,
     },
     material: "wood",
-    height: 200,
-    width: 100,
+    height: 100,
+    width: 300,
     depth: 50,
     doors: 2,
     color: "brown",
